backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

supabase==2.9.0
postgrest==0.17.2

qdrant-client==1.7.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OCR & Processing
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
pypdf==4.0.1
langdetect==1.0.9
openai==1.12.0
